rename typing.py to typehints.py
