+++ audformat/core/index.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 
 from audformat.core import define
-from audformat.core.typing import (
+from audformat.core.typehints import (
     Files,
     Timestamps,
 )
