+++ audformat/core/typehints.py
